Rename directory and change link in header
