+++ src/App/App.css
@@ -7,7 +7,7 @@ font-family: 'Rock Salt', cursive; */
   padding: 0;
 }
 
-.header-link {
+header {
   padding: 10px 30px 0;
   display: flex;
   align-items: center;
