+++ src/NoRoute/NoRoute.css
@@ -7,7 +7,7 @@
 }
 
 .four-oh-four {
-    /*background: url("../images/splat.png");*/
+    /*background: url("../assets/splat.png");*/
     width: 90vw;
     padding-top: -40px;
 }
